feat: tighten loop further
This loop was originally made slower because I thought iterating over the buffs was the issue. The issue was Rasial and how I was detecting the buff's bar. This should significantly speed up how fast the overlay updates from when buffs first appear - from half a tick to a quarter of a tick.

I plan to later introduce a setting to allow players to adjust this as I'm not really sure how intensive this is (if it even is at all...) and so it may be possible to speed it up further or if it starts causing people's fans to turn on they can slow it back down to 300.

"Works on my machine" at least.