feat: Adjust interface search speed
I recommend putting this number as low as you can without encountering performance issues. I have a decent computer so I don't want to presume 10ms is fine for everyone just because it works fine for me.

The initial speed of 125ms is sort of conservative. Not as conservative as the original 300ms though.