feat: Detect t95 oh on very 1st open
On the very first open of the app (or after any full settings resets) if the user has a T95 offhand in their inventory or equipped the app will automatically set the offhand to t95.

I'll get feedback about whether it should run on any open of the app - but I don't want to override user settings if someone has T95 but uses T90 for slayer or something.